Move filtering and deduplicating logic to better spot so extension developers aren't affected
